Prepare to fix contenView top

+++ Application/YandexTestTask/Models/CandlesModel.swift
@@ -8,10 +8,10 @@
 import Foundation
 
 struct CandlesModel: Decodable {
-    let o: [Float]?
-    let h: [Float]?
-    let l: [Float]?
-    let c: [Float]?
+    let o: [Double]?
+    let h: [Double]?
+    let l: [Double]?
+    let c: [Double]?
     let v: [Int]?
     let t: [UInt]?
     let s: String
